move utils into scenario/utils
