+++ scenario/GovernanceScenario.ts
@@ -2,11 +2,11 @@ import { scenario } from './context/CometContext';
 import { expect } from 'chai';
 import { BigNumberish, constants, Contract, EventFilter, utils } from 'ethers';
 import { COMP_WHALES } from "../src/deploy";
-import { impersonateAddress } from '../plugins/scenario/World';
+import { impersonateAddress, setNextBaseFeeToZero, setNextBlockTimestamp } from '../plugins/scenario/utils';
 import { importContract } from '../plugins/deployment_manager/Import';
 import { DeploymentManager } from '../plugins/deployment_manager';
 import { SignerWithAddress } from '@nomiclabs/hardhat-ethers/signers';
-import { isBridgedDeployment, setNextBaseFeeToZero, setNextBlockTimestamp, fastGovernanceExecute  } from './utils';
+import { isBridgedDeployment, fastGovernanceExecute } from './utils';
 
 const FX_ROOT_GOERLI = '0x3d1d3E34f7fB6D26245E6640E1c50710eFFf15bA';
 const STATE_SENDER = '0xeaa852323826c71cd7920c3b4c007184234c3945';
