+++ scenario/context/CometContext.ts
@@ -30,7 +30,8 @@ import { sourceTokens } from '../../plugins/scenario/utils/TokenSourcer';
 import { ProtocolConfiguration, deployComet, COMP_WHALES } from '../../src/deploy';
 import { AddressLike, getAddressFromNumber, resolveAddress } from './Address';
 import { Requirements } from '../constraints/Requirements';
-import { executeOpenProposal, fastGovernanceExecute, mineBlocks, setNextBaseFeeToZero, setNextBlockTimestamp } from '../utils';
+import { executeOpenProposal, fastGovernanceExecute,  } from '../utils';
+import { mineBlocks, setNextBaseFeeToZero, setNextBlockTimestamp } from '../../plugins/scenario/utils';
 
 export type ActorMap = { [name: string]: CometActor };
 export type AssetMap = { [name: string]: CometAsset };
